+++ CHANGELOG.md
@@ -33,6 +33,7 @@
 ### Added
 - Similar artists: new entries to create a similar artists database in JSON equal to the one found at [Search by Distance-SMP](https://github.com/regorxxx/Search-by-Distance-SMP) and tag files with the 10 most similar artists into 'SIMILAR ARTISTS LISTENBRAINZ' tag.
 - YouTube: added support for 'MUSICBRAINZ_ALBUMARTISTID' and 'MUSICBRAINZ_ARTISTID' tags if available for user recommendations.
+- Tags: similar artist tags may now be remapped (globally) at '[FOOBAR_PROFILE]\js_data\presets\global\globTags.json'.
 ### Changed
 - User recommendations: changed the logic to just retrieve and use one of the 10 most similar users, instead of relying on an specific similarity threshold.
 - Similar artists: in case less than 5 artists are retrieved, it will retry the lookup with a different algorithm which usually retrieves more results.